+++ packages/gatsby-plugin-orion-core/gatsby-browser.js
@@ -1,24 +1,9 @@
-import React, { useEffect, useState } from 'react'
-import awsConfig from './utils/aws-exports'
-import AuthWrapper from './src/components/AuthWrapper'
-import { ClientContext } from 'graphql-hooks'
-import { initGraphQLClient, makeGraphQLClient } from './utils/graphql'
+import React, { useEffect } from 'react'
 import { loadCSS } from 'fg-loadcss'
+import AuthWrapper from './src/components/AuthWrapper'
+import GraphQLProvider from './src/components/GraphQLProvider'
 
-const client = makeGraphQLClient(process.env.GATSBY_GRAPHQL_API)
-
-const AuthInitWrapper = ({ element }) => {
-  const [isAuthInitialized, setIsAuthInitialized] = useState(false)
-
-  useEffect(() => {
-    const init = async () => {
-      await initGraphQLClient(client, awsConfig)
-      setIsAuthInitialized(true)
-    }
-
-    init()
-  }, [])
-
+function FontAwesomeLoader({ children }) {
   useEffect(() => {
     loadCSS(
       'https://use.fontawesome.com/releases/v5.12.1/css/all.css',
@@ -26,13 +11,13 @@ const AuthInitWrapper = ({ element }) => {
     )
   }, [])
 
-  return (
-    <ClientContext.Provider value={client}>
-      <AuthWrapper isAuthInitialized={isAuthInitialized}>{element}</AuthWrapper>
-    </ClientContext.Provider>
-  )
+  return children
 }
 
 export const wrapRootElement = ({ element }) => (
-  <AuthInitWrapper client={client} element={element} />
+  <FontAwesomeLoader>
+    <AuthWrapper>
+      <GraphQLProvider>{element}</GraphQLProvider>
+    </AuthWrapper>
+  </FontAwesomeLoader>
 )