+++ packages/gatsby-plugin-orion-view/src/components/UserRegistration/index.js
@@ -1,6 +1,6 @@
 import React, { useState } from 'react'
 import { Grid, Button, Typography } from '@material-ui/core'
-import { Auth } from 'aws-amplify'
+import { Auth } from 'gatsby-plugin-orion-core/src/utils/amplify'
 
 import { Form } from 'gatsby-plugin-orion-core'